requirements.txt
aiosmtplib==5.1.0
alembic==1.18.4
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
async-timeout==5.0.1
asyncpg==0.31.0
attrs==25.4.0
bcrypt==4.3.0
beautifulsoup4==4.14.3
billiard==4.2.4
blinker==1.9.0
blis==1.3.3
bs4==0.0.2
cachetools==6.2.6
catalogue==2.0.10
celery==5.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
dateparser==1.2.2
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.129.0
filelock==3.20.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
murmurhash==1.0.15
narwhals==2.17.0
networkx==3.6.1
numpy==2.4.2
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pdfminer.six==20251107
pdfplumber==0.11.8
phonenumbers==9.0.21
pillow==12.1.1
plotly==6.5.2
preshed==3.0.12
prompt_toolkit==3.0.52
protobuf==6.33.5
psycopg2==2.9.11
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pypdfium2==5.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
redis==7.2.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
selenium==4.39.0
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
srsly==2.5.2
starlette==0.52.1
streamlit==1.54.0
streamlit-autorefresh==1.0.1
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
trio==0.32.0
trio-websocket==0.12.2
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
weasel==0.4.3
websocket-client==1.9.0
websockets==16.0
wheel==0.45.1
wrapt==2.0.1
wsproto==1.3.2
